─ requirements.txt
langchain
langchain-openai
langgraph
tavily-python
python-dotenv